Update BD variables names.
************************************************************
WARNINGS (3)
blackduck.hub.password: This property is changing. Please use --blackduck.password in the future. It will cause failure in 6.0.0 and be removed in 7.0.0.
blackduck.hub.url: This property is changing. Please use --blackduck.url in the future. It will cause failure in 6.0.0 and be removed in 7.0.0.
blackduck.hub.username: This property is changing. Please use --blackduck.username in the future. It will cause failure in 6.0.0 and be removed in 7.0.0.
************************************************************